+++ src/Simpletools/Db/Cassandra/Async.php
@@ -1,16 +1,19 @@
 <?php
 
 namespace Simpletools\Db\Cassandra;
+use Simpletools\Db\Cassandra\Type\Uuid;
 use Simpletools\Db\Replicator;
 
 class Async
 {
 	protected $_queries = array();
+	protected $_queriesCache = array(); //[$query, $options, $retried]
 	protected $_client;
 	protected $_runOnBatchSize = 10000; //0 - only manual batch run()
 	protected $_requestTimeout = 5;
 	protected $_replicationQuery = [];
 	protected $_replication = true;
+	protected $_retryPolicy = 'reconnect'; //(reconnect | fallthrough | silence )
 
 	public function __construct($batchSize = 10000, $requestTimeout = 5, $client = null)
 	{
@@ -32,6 +35,17 @@ class Async
 		$this->_client->connect();
 	}
 
+	/**
+	 * @param string $policy (reconnect | fallthrough | silence )
+	 */
+	public function setRetryPolicy(string $policy)
+	{
+		if(!in_array($policy,['reconnect', 'fallthrough', 'silence']))
+			throw  new Exception('Unknow retryPolicy: '.$policy.' Use: (reconnect | fallthrough | silence)');
+
+		$this->_retryPolicy = $policy;
+	}
+
 	public function disableReplication()
 	{
 		$this->_replication = false;
@@ -39,6 +53,7 @@ class Async
 
 	public function add($query)
 	{
+		$streamId = (string)(new Uuid());
 		if($query instanceof Batch)
 		{
 			if(!$query->size())
@@ -52,14 +67,15 @@ class Async
                 $options['consistency'] = $consistency;
             }
 
-			$batchId = uniqid();
-			$this->_queries[$batchId] = $this->_client->connector()->executeAsync($query->getBatch(),$options);
+			$rawQuery = $query->getBatch();
+			$this->_queries[$streamId] = $this->_client->connector()->executeAsync($rawQuery,$options);
+			if($this->_retryPolicy =='reconnect')
+				$this->_queriesCache[$streamId] = [$rawQuery,$options, 0];
 
 			if($query->isReplication())
 			{
-
 				$repQuery = $query->getReplicationQuery();
-				$this->_replicationQuery[$batchId] = (object)[
+				$this->_replicationQuery[$streamId] = (object)[
 					'data' => $repQuery->data,
 					'type' => 'bulk',
 					'dest' => $repQuery->keyspace.'.'.$repQuery->table
@@ -91,7 +107,6 @@ class Async
 		if($this->_replication)
 		{
 			$rawQuery = $query->getRawQuery();
-			$callId = uniqid();
 
 			$type = false;
 			if($rawQuery['type'] == 'INSERT') $type ='write';
@@ -103,7 +118,7 @@ class Async
 				if($data = $query->getRawQueryData($rawQuery))
 				{
 					$this->_replication = true;
-					$this->_replicationQuery[$callId] = (object)[
+					$this->_replicationQuery[$streamId] = (object)[
 						'type' =>  $type,
 						'data' => $data,
 						'dest' => $rawQuery['db'].'.'.$rawQuery['table']
@@ -117,7 +132,11 @@ class Async
 			if($consistency!==null)
                 $options['consistency'] = $consistency;
 
-			$this->_queries[$callId] = $this->_client->connector()->executeAsync($query['preparedQuery'],$options);
+			$this->_queries[$streamId] = $this->_client->connector()->executeAsync($query['preparedQuery'],$options);
+			if($this->_retryPolicy =='reconnect')
+			{
+				$this->_queriesCache[$streamId] = [$query['preparedQuery'],$options, 0];
+			}
 		}
 		else
 		{
@@ -127,7 +146,11 @@ class Async
             if($consistency!==null)
                 $options['consistency'] = $consistency;
 
-			$this->_queries[] = $this->_client->connector()->executeAsync($query['preparedQuery'],$options);
+			$this->_queries[$streamId] = $this->_client->connector()->executeAsync($query['preparedQuery'],$options);
+			if($this->_retryPolicy =='reconnect')
+			{
+				$this->_queriesCache[$streamId] = [$query['preparedQuery'],$options,0];
+			}
 		}
 
 		if($this->_runOnBatchSize && count($this->_queries) >= $this->_runOnBatchSize)
@@ -145,28 +168,96 @@ class Async
 			throw new Exception("Empty async bulk",400);
 		}
 
-		if($this->_replication)
-		{
-			foreach ($this->_queries as $queryId => $future)
-			{
-				$future->get($this->_requestTimeout);
-				unset($this->_queries[$queryId]);
 
-				if(isset($this->_replicationQuery[$queryId]))
+		if($this->_replication && $this->_retryPolicy=='reconnect')
+		{
+			foreach ($this->_queries as $streamId => $future)
+			{
+				try{
+					$future->get($this->_requestTimeout);
+					unset($this->_queries[$streamId]);
+					unset($this->_queriesCache[$streamId]);
+				}catch (\Exception $e)
 				{
-					Replicator::trigger('cassandra://'.$this->_replicationQuery[$queryId]->type.'@'.$this->_replicationQuery[$queryId]->dest, $this->_replicationQuery[$queryId]->data);
-					unset($this->_replicationQuery[$queryId]);
+					if (($e->getCode() == 16777230 || $e->getCode() == '16777225') && isset($this->_queriesCache[$streamId]) && $this->_queriesCache[$streamId][2] < 3)
+					{
+						$this->_queriesCache[$streamId][2]++;
+						$this->_queries[$streamId] = $this->_client->connector()->executeAsync($this->_queriesCache[$streamId][0],$this->_queriesCache[$streamId][1]);
+					}
+					else
+						throw $e;
 				}
 			}
 		}
-		else
+		elseif($this->_replication && $this->_retryPolicy=='fallthrough')
 		{
-			foreach ($this->_queries as $queryId => $future)
+			foreach ($this->_queries as $streamId => $future)
 			{
 				$future->get($this->_requestTimeout);
-				unset($this->_queries[$queryId]);
+				unset($this->_queries[$streamId]);
+
+				if(isset($this->_replicationQuery[$streamId]))
+				{
+					Replicator::trigger('cassandra://'.$this->_replicationQuery[$streamId]->type.'@'.$this->_replicationQuery[$queryId]->dest, $this->_replicationQuery[$queryId]->data);
+					unset($this->_replicationQuery[$streamId]);
+				}
 			}
 		}
+		elseif($this->_replication && $this->_retryPolicy=='silence')
+		{
+			foreach ($this->_queries as $streamId => $future)
+			{
+				try{
+					$future->get($this->_requestTimeout);
+					if(isset($this->_replicationQuery[$streamId]))
+					{
+						Replicator::trigger('cassandra://'.$this->_replicationQuery[$streamId]->type.'@'.$this->_replicationQuery[$queryId]->dest, $this->_replicationQuery[$queryId]->data);
+						unset($this->_replicationQuery[$streamId]);
+					}
+				} catch (\Exception $e){}
+				unset($this->_queries[$streamId]);
+			}
+		}
+		elseif($this->_retryPolicy=='reconnect')
+		{
+			foreach ($this->_queries as $streamId => $future)
+			{
+				try{
+				$future->get($this->_requestTimeout);
+					unset($this->_queries[$streamId]);
+					unset($this->_queriesCache[$streamId]);
+				}catch (\Exception $e)
+				{
+					if (($e->getCode() == 16777230 || $e->getCode() == '16777225') && isset($this->_queriesCache[$streamId]) && $this->_queriesCache[$streamId][2] < 3)
+					{
+						$this->_queriesCache[$streamId][2]++;
+						$this->_queries[$streamId] = $this->_client->connector()->executeAsync($this->_queriesCache[$streamId][0],$this->_queriesCache[$streamId][1]);
+					}
+					else
+							throw $e;
+				}
+			}
+		}
+		elseif($this->_retryPolicy=='fallthrough')
+		{
+			foreach ($this->_queries as $streamId => $future)
+			{
+				$future->get($this->_requestTimeout);
+				unset($this->_queries[$streamId]);
+			}
+		}
+		elseif($this->_retryPolicy=='silence')
+		{
+			foreach ($this->_queries as $streamId => $future)
+			{
+				try{
+					$future->get($this->_requestTimeout);
+				} catch (\Exception $e){}
+				unset($this->_queries[$streamId]);
+			}
+		}
+
+		if($this->_queries) $this->run();
 
 		$this->reset();
 
@@ -185,6 +276,7 @@ class Async
 	{
 		$this->_queries   = array();
 		$this->_replicationQuery =  [];
+		$this->_queriesCache = [];
 
 		return $this;
 	}